+++ src/api/repositories/ApiRepository.ts
@@ -546,8 +546,14 @@ export class ApiRepository {
     module?: ModuleType;
     inspectionScope?: InspectionScope;
     teamId?: string;
+    contractId?: string;
     status?: InspectionStatus;
     osNumber?: string;
+    service?: string;
+    executionFrom?: string;
+    executionTo?: string;
+    inspectionFrom?: string;
+    inspectionTo?: string;
     investmentWorkId?: string;
     page?: number;
     limit?: number;