+++ src/repositories/IAppRepository.ts
@@ -243,8 +243,14 @@ export interface IAppRepository {
     module?: ModuleType;
     inspectionScope?: InspectionScope;
     teamId?: string;
+    contractId?: string;
     status?: InspectionStatus;
     osNumber?: string;
+    service?: string;
+    executionFrom?: string;
+    executionTo?: string;
+    inspectionFrom?: string;
+    inspectionTo?: string;
     investmentWorkId?: string;
     page?: number;
     limit?: number;